Use buildah to build images
Added alternative builder for project images: buildah
https://github.com/containers/buildah

Considering that some of our plugins use CRI-O runtime it could be
a good idea to get rid of docker as a builder. It should allow us
not to run docker daemon at all, even for build purposes.

Kubernetes also goes this way encouraging users to switch to CRI
runtimes (CRI-O and containerd), so having non-docker builds supported
looks good from this perspective too.

+++ demo/build-image.sh
@@ -1,6 +1,7 @@
-#!/bin/sh -xe
+#!/bin/sh -e
 
 IMG=$1
+BUILDER=$2
 
 if [ -z "$IMG" ]; then
     (>&2 echo "Usage: $0 <image directory>")
@@ -15,5 +16,13 @@ fi
 CWD=`dirname $0`
 TAG=`git rev-parse HEAD`
 
-docker build --rm -t ${IMG}:${TAG} "$CWD/$IMG/"
-docker tag ${IMG}:${TAG} ${IMG}:devel
+if [ -z "$BUILDER" -o "$BUILDER" = 'docker' ] ; then
+    docker build --rm -t ${IMG}:${TAG} "$CWD/$IMG/"
+    docker tag ${IMG}:${TAG} ${IMG}:devel
+elif [ "$BUILDER" = 'buildah' ] ; then
+    buildah bud -t ${IMG}:${TAG} "$CWD/$IMG/"
+    buildah tag ${IMG}:${TAG} ${IMG}:devel
+else
+    (>&2 echo "Unknown builder $BUILDER")
+    exit 1
+fi
