+++ deployments/fpga_plugin/build-initcontainer-image.sh
@@ -1,5 +1,7 @@
 #!/bin/sh -e
 
+BUILDER=$1
+
 # check if DCP tarball is present
 DCP_TARBALL=a10_gx_pac_ias_1_1_pv_rte_installer.tar.gz
 if [ ! -f ${DCP_TARBALL} ] ; then
@@ -18,5 +20,13 @@ WORKSPACE=`realpath .`
 IMG="intel-fpga-initcontainer"
 TAG=$(git rev-parse HEAD)
 
-docker build --rm -t $IMG:devel -f $WORKSPACE/$IMG.Dockerfile $WORKSPACE
-docker tag $IMG:devel $IMG:$TAG
+if [ -z "${BUILDER}" -o "${BUILDER}" = 'docker' ] ; then
+    docker build --rm -t $IMG:devel -f $WORKSPACE/$IMG.Dockerfile $WORKSPACE
+    docker tag $IMG:devel $IMG:$TAG
+elif [ "${BUILDER}" = 'buildah' ] ; then
+    buildah bud -t $IMG:devel -f $WORKSPACE/$IMG.Dockerfile --layers $WORKSPACE
+    buildah tag $IMG:devel $IMG:$TAG
+else
+    (>&2 echo "Unknown builder ${BUILDER}")
+    exit 1
+fi
